+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [0.7.1]
 ### Changed
  - Removed issue with brutal crashes causing data loss on database
+ - Fixed issue with periodic reset (move to other versions too)
 
 ## [0.7.0]
 ### Added